# docs/swagger.yaml
definitions:
  AuthResponse:
    properties:
      token:
        type: string
    type: object
...
  ErrorResponse:
    properties:
      error:
        properties:
          code:
            enum:
            - FORBIDDEN
            - UNAUTHORIZED
            - INVALID_CREDENTIALS
            - INVALID_REQUEST
            - INTERNAL_ERROR
            - FORM_NOT_FOUND
            - FORM_ALREADY_REJECTED
            - FORM_ALREADY_APPROVED
            - NO_AVAILABLE_EXECUTORS
            - USER_NOT_FOUND
            - USER_NOT_ACTIVE
            - USER_ALREADY_EXISTS
            - DOCUMENT_NOT_FOUND
            example: ERROR_CODE
            type: string
          message:
            type: string
        required:
        - code
        - message
        type: object
    required:
    - error
    type: object
...
  LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  RegisterRequest:
    properties:
      email:
        type: string
      firstName:
        minLength: 2
        type: string
      lastName:
        minLength: 2
        type: string
      password:
        minLength: 4
        type: string
      position:
        minLength: 2
        type: string
      role:
        enum:
        - employee
        - hr
        type: string
    required:
    - email
    - firstName
    - lastName
    - password
    - position
    - role
    type: object
...
info:
  contact: {}
  description: RESTful API для приложения HRMate. Управление пользователями, авторизацией
    и формами.
  title: HRMate API
  version: "1.0"
paths:
...
  /login:
    post:
      consumes:
      - application/json
      description: Authenticate a user and return an auth token
      parameters:
      - description: User credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: User login
      tags:
      - Auth
...
  /register:
    post:
      consumes:
      - application/json
      description: Register a new user and return an auth token
      parameters:
      - description: User registration details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Register a new user
      tags:
      - Auth
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"